// src/app/admin/products/products.types.ts
import { Category } from '@/src/app/admin/categories/categories.types';
export type ProductWithCategory = {
    category: Category;
    created_at: string;
    heroImage: string;
    id: number;
    imagesUrl: string[];
    maxQuantity: number;
    price: number | null;
    slug: string;
    title: string;
};

export type ProductsWithCategoriesResponse = ProductWithCategory[];

export type UpdateProductSchema = {
    category: number;
    heroImage: string;
    imagesUrl: string[];
    maxQuantity: number;
    price: number;
    slug: string;
    title: string;
};