// src/app/admin/products/schema.ts
import { z } from 'zod'
export const createOrUpdateProductsSchema = z.object({
    title: z.string().min(1, { message: 'title is required' }),
    price: z.string().min(1, { message: 'price is required' }),
    maxQuantity: z.string().min(1, { message: 'maxquantity is required' }),
    category: z.string().min(1, { message: 'category is required' }),
    heroImage: z.any().refine(file => file.length === 1, 'heroImage is required'),
    images: z.any().refine((files: FileList | null) => files instanceof FileList && files.length > 0,
        { message: 'at least one image is required' })
        .transform((files: FileList | null) => (files ? Array.from(files) : [])),
    intent: z.enum(['create', 'update'], {
        message: 'intent must be either create or update',
    })
        .optional(),
    slug: z.string().optional(),
})
export type CreateOrUpdateProductSchema = z.infer<
    typeof createOrUpdateProductsSchema>
export const createProductSchemaServer = z.object({
    title: z.string().min(1, { message: 'title is required' }),
    price: z.number().min(1, { message: 'price is required' }),
    maxQuantity: z.number().min(1, { message: 'maxquantity is required' }),
    category: z.number().min(1, { message: 'category is required' }),
    heroImage: z.string().url({ message: 'hero image is required' }),
    images: z.array(z.string().url({ message: 'images are required' })),
})
export type CreateProductSchemaServer = z.infer<
    typeof createProductSchemaServer>